requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
gunicorn==21.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
firebase-admin==6.6.0
